requirements.txt
aiogram==3.13.1
fastapi==0.115.6
python-dotenv==1.0.1
uvicorn==0.32.1